# inventarioBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Inventario API",
        "description": "Sistema para gestionar inventario para empresas en crecimiento",
        "contact": {
            "name": "API Support",
            "email": "support@inventario.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Obtiene una lista de todos los productos",
                "operationId": "389c66353d798efd6a7dfda57d99edd9",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "description": "Crea un nuevo producto en el inventario",
                "operationId": "3e5e3b44905c09b5642f99a1abd861c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "cantidad_disponible",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Producto Ejemplo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del producto",
                                        "nullable": true
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "cantidad_disponible": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Electrónica",
                                        "nullable": true
                                    },
                                    "proveedor": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "codigoProducto": {
                                        "type": "string",
                                        "example": "PROD-001",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos no válidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar producto",
                "description": "Obtiene la información de un producto específico",
                "operationId": "56daf8c60b3db5519da35f337589a475",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "description": "Actualiza la información de un producto existente",
                "operationId": "a624accb933639b2dc0d87064011accc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Producto Ejemplo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del producto",
                                        "nullable": true
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "cantidad_disponible": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Electrónica",
                                        "nullable": true
                                    },
                                    "proveedor": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "codigoProducto": {
                                        "type": "string",
                                        "example": "PROD-001",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Elimina un producto del inventario",
                "operationId": "9814c4f94338a6b85ba4ecdac9e5afbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El producto se ha eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/productos/{id}/cambiar-estado": {
            "patch": {
                "tags": [
                    "Productos"
                ],
                "summary": "Cambiar estado del producto",
                "description": "Alterna el estado (activo/inactivo) de un producto",
                "operationId": "3ef55d6023c373b4e4e0cde7b74d6447",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/productos/buscar": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Buscar productos por nombre",
                "description": "Busca productos por nombre (búsqueda parcial)",
                "operationId": "896813d9a79862ce74cde434a3066c8d",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/productos/ordenar": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Ordenar productos por precio",
                "description": "Obtiene productos ordenados por precio",
                "operationId": "60c47d03d8ef5acb8747a8a96ac2b8a5",
                "parameters": [
                    {
                        "name": "orden",
                        "in": "query",
                        "description": "Orden ascendente o descendente",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos ordenados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Producto": {
                "properties": {
                    "IdProducto": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Producto Ejemplo"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción del producto",
                        "nullable": true
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "cantidad_disponible": {
                        "type": "integer",
                        "example": 50
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Electrónica",
                        "nullable": true
                    },
                    "proveedor": {
                        "type": "integer",
                        "nullable": true
                    },
                    "codigoProducto": {
                        "type": "string",
                        "example": "PROD-001",
                        "nullable": true
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "name": "bearerAuth",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}